condensed finds.js pages
